Split JavaApiIntegrationTest tests for better maintainability
